+++ src/app/core/components/app-sidebar/app-sidebar.proxy.ts
@@ -3,19 +3,15 @@ import { Store } from '@ngrx/store';
 import { AppApi } from '@api/app.api';
 import { EchoesState } from '@store/reducers';
 
-import { getSidebarCollapsed } from '@store/app-layout';
+import { getSidebarCollapsed } from '@store/app-core';
 import * as PlayerSearch from '@store/player-search';
-import * as AppLayout from '@store/app-layout';
 
 @Injectable()
 export class AppSidebarProxy {
-
   sidebarCollapsed$ = this.store.select(getSidebarCollapsed);
   searchType$ = this.store.select(PlayerSearch.getSearchType);
 
-  constructor(
-    private store: Store<EchoesState>,
-    private appApi: AppApi) { }
+  constructor(private store: Store<EchoesState>, private appApi: AppApi) {}
 
   toggleSidebar() {
     this.appApi.toggleSidebar();