+++ src/app/core/components/index.ts
@@ -3,8 +3,10 @@ import { AppNavigatorModule } from './app-navigator';
 import { NowPlayingModule } from './now-playing';
 import { AppBrandModule } from './app-brand';
 import { AppSidebarModule } from './app-sidebar';
+import { AppErrorHandlerModule } from './app-error-handler';
 
 export const APP_CORE_MODULES = [
   AppPlayerModule,
-  AppSidebarModule
+  AppSidebarModule,
+  AppErrorHandlerModule
 ];