+++ src/app/core/effects/router.effects.ts
@@ -1,31 +1,32 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { Location } from '@angular/common';
-import { Effect, Actions } from '@ngrx/effects';
+import { Effect, Actions, ofType } from '@ngrx/effects';
 import * as fromRouterActions from '@store/router-store';
 import { map, tap } from 'rxjs/operators';
 
 @Injectable()
 export class RouterEffects {
   @Effect({ dispatch: false })
-  navigate$ = this.actions$
-    .ofType(fromRouterActions.GO)
-    .pipe(
-      map((action: fromRouterActions.Go) => action.payload),
-      tap(({ path, query: queryParams, extras }) =>
-        this.router.navigate(path, { queryParams, ...extras })
-      )
-    );
+  navigate$ = this.actions$.pipe(
+    ofType(fromRouterActions.GO),
+    map((action: fromRouterActions.Go) => action.payload),
+    tap(({ path, query: queryParams, extras }) =>
+      this.router.navigate(path, { queryParams, ...extras })
+    )
+  );
 
   @Effect({ dispatch: false })
-  navigateBack$ = this.actions$
-    .ofType(fromRouterActions.BACK)
-    .pipe(tap(() => this.location.back()));
+  navigateBack$ = this.actions$.pipe(
+    ofType(fromRouterActions.BACK),
+    tap(() => this.location.back())
+  );
 
   @Effect({ dispatch: false })
-  navigateForward$ = this.actions$
-    .ofType(fromRouterActions.FORWARD)
-    .pipe(tap(() => this.location.forward()));
+  navigateForward$ = this.actions$.pipe(
+    ofType(fromRouterActions.FORWARD),
+    tap(() => this.location.forward())
+  );
 
   constructor(
     private actions$: Actions,