+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v 4.2.0 (2019/03/17)
+
+* [UPGRADE] - upgraded to Angular 7
+* [NEW] - added toast for errors (error handler)
+* [NEW] - add http interceptor with error handling (offline error toast)
+
 ## v 4.1.1 (2019/02/26) (codename: maya-1)
 
 * [NEW] - icon overhaul - now using line-awesome